refactor: Answer 코드정리
