+++ baseball/src/main/java/com/flab/baseball/application/GameStartProcessor.java
@@ -1,8 +1,8 @@
 package com.flab.baseball.application;
 
 import com.flab.baseball.application.data.GameStartData;
-import com.flab.baseball.domain.Answer;
-import com.flab.baseball.domain.RandomAnswerGenerator;
+import com.flab.baseball.domain.answer.Answer;
+import com.flab.baseball.domain.answer.RandomAnswerGenerator;
 import com.flab.baseball.domain.Room;
 import com.flab.baseball.domain.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +20,7 @@ public class GameStartProcessor {
 				new RandomAnswerGenerator()
 			)
 		);
+
 		Room savedRoom = roomRepository.persist(room);
 		return new GameStartData(savedRoom.getId());
 	}
